pyproject.toml
[project]
name = "neuro-rag"
version = "0.1.0"
description = "神經內科醫療 RAG 系統 - 榮民醫院門診問答系統"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # === LangChain 核心 ===
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",

    # === LLM 提供商 ===
    "langchain-openai>=0.2.0",       # OpenAI GPT
    "langchain-anthropic>=0.2.0",    # Claude
    # "langchain-ollama>=0.2.0",     # 本地部署時取消註解

    # === Embedding ===
    "sentence-transformers>=3.0.0",  # 載入 BGE 模型
    "langchain-huggingface>=0.1.0",  # HuggingFace 整合

    # === 向量資料庫 ===
    "chromadb>=0.5.0",               # 開發用向量 DB
    "langchain-chroma>=0.1.0",       # Chroma 整合

    # === 爬蟲相關 ===
    "httpx>=0.27.0",                 # HTTP 請求
    "beautifulsoup4>=4.12.0",        # HTML 解析
    "lxml>=5.0.0",                   # 解析器後端
    # "playwright>=1.40.0",          # 動態網頁（需要時安裝）

    # === 文件處理 ===
    "pymupdf>=1.24.0",               # PDF 解析

    # === 設定與工具 ===
    "pydantic>=2.0.0",               # 資料驗證
    "pydantic-settings>=2.0.0",      # 設定管理
    "pyyaml>=6.0.0",                 # YAML 解析
    "python-dotenv>=1.0.0",          # 環境變數
    "typer>=0.12.0",                 # CLI 介面
    "rich>=13.0.0",                  # 美化輸出
    "loguru>=0.7.0",                 # 日誌
]

[project.optional-dependencies]
# 開發工具
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",                   # Linter + Formatter
    "mypy>=1.11.0",                  # 型別檢查
    "ipython>=8.0.0",                # 互動式開發
    "matplotlib>=3.0.0",             # 視覺化
    "scikit-learn>=1.0.0",           # 降維 (PCA)
]

# 本地 LLM 部署
local-llm = [
    "langchain-ollama>=0.2.0",
    # vLLM 需要特殊安裝，不放這裡
]

# 生產部署（大規模向量 DB）
production = [
    "pymilvus>=2.4.0",               # Milvus 向量資料庫
]

# Jupyter 開發
notebook = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.scripts]
# CLI 命令
neuro-rag = "src.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# === 開發工具設定 ===

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # 行長度（由 formatter 處理）
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
